refactor(game): make registry use short as block rid

+++ server/src/main/kotlin/com/github/bedrockecs/server/game/registry/BlockRegistry.kt
@@ -19,12 +19,12 @@ interface BlockRegistry {
     /**
      * get [BlockTypeComponent] associated with this runtimeId
      */
-    fun typeFor(rid: Int): BlockTypeComponent?
+    fun typeFor(rid: Short): BlockTypeComponent?
 
     /**
      * the runtime id of block of this type, used in serialization etc......
      */
-    fun runtimeIDFor(type: BlockTypeComponent): Int?
+    fun runtimeIDFor(type: BlockTypeComponent): Short?
 
     // Default BlockComponents //
